// src/app/models/Supervisor.ts
export class Supervisor {

    id: string | undefined;
    nombre: string | undefined;
    apellidos: string| undefined;
    genero: string | undefined;
    telefono: string| undefined;
    email: string| undefined;
    contraseña: string| undefined;
    direccion: string| undefined;
    id_rol: number| undefined;
    created_at: string| undefined;
}